requirements.txt
imaplib2
email
pandas
python-dotenv